save and restore main for islands tests
